requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
PyMySQL==1.1.1
python-multipart==0.0.9
google-api-python-client==2.169.0
google-auth==2.40.1
pandas==2.2.2
openpyxl==3.1.5
